backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==3.0.2
pillow==11.0.0
pycparser==2.22
python-dotenv==1.0.1
python-jwt==4.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
